mlx5: fix clippy warnings and add type aliases for QP handles
- Fix various clippy warnings:
  - Remove unused imports
  - Fix unnecessary_cast in device.rs
  - Fix redundant_closure in dc/mod.rs
  - Use unwrap_or_default() instead of unwrap_or_else(Weak::new)
  - Fix collapsible_if in ud/builder.rs
  - Fix unused_unsafe in barrier.rs macro
  - Fix unused_mut and unused_variables in test files
  - Fix unused Result in wraparound_tests.rs
  - Add #[allow(dead_code)] for test utilities in common.rs
  - Fix loop variable indexing in pingpong benchmark

- Add type aliases for QP handles to fix type_complexity warnings:
  - RcQpIbHandle, RcQpIbWithSrqHandle, RcQpRoCEHandle, RcQpRoCEWithSrqHandle
  - UdQpIbHandle, UdQpIbWithSrqHandle, UdQpRoCEHandle, UdQpRoCEWithSrqHandle

- Add #[allow(clippy::type_complexity)] for benchmark setup functions
  (impl Fn return types cannot be easily aliased)

+++ mlx5/src/device.rs
@@ -352,9 +352,9 @@ impl Context {
             }
             let attr = attr.assume_init();
             Some(PciAtomicCaps {
-                fetch_add: attr.pci_atomic_caps.fetch_add as u16,
-                swap: attr.pci_atomic_caps.swap as u16,
-                compare_swap: attr.pci_atomic_caps.compare_swap as u16,
+                fetch_add: attr.pci_atomic_caps.fetch_add,
+                swap: attr.pci_atomic_caps.swap,
+                compare_swap: attr.pci_atomic_caps.compare_swap,
             })
         }
     }
